Rename gradio.py to app.py
